requirements.txt
# ---- Core ML / Deep Learning ----
torch==2.7.0
torchvision==0.22.0

# ---- Math & Scientific Libraries ----
numpy==2.1.3
scipy==1.14.1
scikit-learn==1.5.2
matplotlib==3.9.2
tqdm==4.67.0
opencv-python==4.10.0.84
Pillow==11.0.0

# ---- Web Framework ----
flask==3.0.3
